requirements.txt
einops
transformers>=4.25.1
safetensors>=0.3.0
aiohttp
accelerate
pyyaml
Pillow
scipy
tqdm
psutil
nltk
matplotlib
sentencepiece
together>=0.1.5